+++ common/src/main/java/com/devnemo/nemos/copper/block/CopperPoweredRailBlock.java
@@ -1,0 +1,10 @@
+package com.devnemo.nemos.copper.block;
+
+import net.minecraft.world.level.block.PoweredRailBlock;
+
+public class CopperPoweredRailBlock extends PoweredRailBlock {
+
+    public CopperPoweredRailBlock(Properties properties) {
+        super(properties);
+    }
+}