+++ common/src/main/java/com/devnemo/nemos/copper/item/ModToolMaterials.java
@@ -1,0 +1,10 @@
+package com.devnemo.nemos.copper.item;
+
+import com.devnemo.nemos.copper.tag.ModBlockTags;
+import com.devnemo.nemos.copper.tag.ModItemTags;
+import net.minecraft.world.item.ToolMaterial;
+
+public class ModToolMaterials {
+
+    public static final ToolMaterial COPPER = new ToolMaterial(ModBlockTags.INCORRECT_FOR_COPPER_TOOL, 200, 5.0F, 1.5F, 12, ModItemTags.COPPER_TOOL_MATERIALS);
+}