+++ common/src/main/java/com/nemonotfound/nemos/copper/interfaces/CopperBucketItemStackGetter.java
@@ -1,8 +1,0 @@
-package com.nemonotfound.nemos.copper.interfaces;
-
-import net.minecraft.world.item.ItemStack;
-
-public interface CopperBucketItemStackGetter {
-
-    ItemStack nemosCopper$getCopperBucketItemStack();
-}