+++ fabric/src/main/java/com/devnemo/nemos/copper/NemosCopperClientFabric.java
@@ -1,0 +1,32 @@
+package com.devnemo.nemos.copper;
+
+import com.devnemo.nemos.copper.block.ModBlocks;
+import com.devnemo.nemos.copper.client.renderer.entity.CopperMinecartRenderer;
+import com.devnemo.nemos.copper.world.entity.ModEntityTypes;
+import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
+import net.fabricmc.fabric.api.client.rendering.v1.EntityRendererRegistry;
+import net.minecraft.client.model.geom.ModelLayers;
+import net.minecraft.client.renderer.RenderType;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+
+public class NemosCopperClientFabric implements ClientModInitializer {
+
+    @Override
+    public void onInitializeClient() {
+        EntityRendererRegistry.register(ModEntityTypes.COPPER_MINECART.get(), context -> new CopperMinecartRenderer(context, ModelLayers.MINECART));
+        EntityRendererRegistry.register(ModEntityTypes.COPPER_CHEST_MINECART.get(), context -> new CopperMinecartRenderer(context, ModelLayers.CHEST_MINECART));
+        EntityRendererRegistry.register(ModEntityTypes.COPPER_COMMAND_BLOCK_MINECART.get(), context -> new CopperMinecartRenderer(context, ModelLayers.COMMAND_BLOCK_MINECART));
+        EntityRendererRegistry.register(ModEntityTypes.COPPER_FURNACE_MINECART.get(), context -> new CopperMinecartRenderer(context, ModelLayers.FURNACE_MINECART));
+        EntityRendererRegistry.register(ModEntityTypes.COPPER_HOPPER_MINECART.get(), context -> new CopperMinecartRenderer(context, ModelLayers.HOPPER_MINECART));
+        EntityRendererRegistry.register(ModEntityTypes.COPPER_SPAWNER_MINECART.get(), context -> new CopperMinecartRenderer(context, ModelLayers.SPAWNER_MINECART));
+        EntityRendererRegistry.register(ModEntityTypes.COPPER_TNT_MINECART.get(), context -> new CopperMinecartRenderer(context, ModelLayers.TNT_MINECART));
+        BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.COPPER_LANTERN.get(), RenderType.cutout());
+        BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.COPPER_SOUL_LANTERN.get(), RenderType.cutout());
+        BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.COPPER_CHAIN.get(), RenderType.cutout());
+        BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.COPPER_BARS.get(), RenderType.cutout());
+        BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.COPPER_HOPPER.get(), RenderType.cutoutMipped());
+
+        BlockEntityType.HOPPER.addSupportedBlock(ModBlocks.COPPER_HOPPER.get());
+    }
+}